refactor transaction types

+++ payment-gateway-authorizer/src/main/java/groupB/newbankV5/paymentgateway/connectors/dto/CreditCardInformationDto.java
@@ -7,12 +7,12 @@ public class CreditCardInformationDto {
     private String cardNumber;
     private String expirationDate;
     private String cvv;
-    private BigDecimal amount;
+    private double amount;
 
     public CreditCardInformationDto() {
     }
 
-    public CreditCardInformationDto(String cardNumber, String expirationDate, String cvv, BigDecimal amount) {
+    public CreditCardInformationDto(String cardNumber, String expirationDate, String cvv, double amount) {
         this.cardNumber = cardNumber;
         this.expirationDate = expirationDate;
         this.cvv = cvv;
@@ -53,11 +53,11 @@ public class CreditCardInformationDto {
                 '}';
     }
 
-    public BigDecimal getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
 }
