+++ payment-gateway-confirmation/src/main/java/groupB/newbankV5/paymentgateway/interfaces/IMockBank.java
@@ -3,5 +3,5 @@ package groupB.newbankV5.paymentgateway.interfaces;
 import java.math.BigDecimal;
 
 public interface IMockBank {
-    String reserveFunds(BigDecimal amount, String cardNumber, String expiryDate, String cvv);
+    String reserveFunds(double amount, String cardNumber, String expiryDate, String cvv);
 }
