+++ payment-processor/src/main/java/groupB/newbankV5/paymentprocessor/interfaces/IFundsHandler.java
@@ -3,16 +3,15 @@ package groupB.newbankV5.paymentprocessor.interfaces;
 import groupB.newbankV5.paymentprocessor.connectors.dto.AccountDto;
 import groupB.newbankV5.paymentprocessor.controllers.dto.TransferDto;
 
-import java.math.BigDecimal;
 
 public interface IFundsHandler {
 
-    boolean hasSufficientFunds(AccountDto accountDto, BigDecimal amount);
+    boolean hasSufficientFunds(AccountDto accountDto, double amount);
 
 
-    void deductFunds(long accountId, BigDecimal amount);
+    void deductFunds(long accountId, double amount);
 
-    void depositFund(long accountId, BigDecimal amount);
+    void depositFund(long accountId, double amount);
 
     boolean isFraudulent(TransferDto transaction);
 }
